Update tests directory structure
