+++ contract/test/tools/prepare-test-env-ava.js
