+++ contract/test/unitTests/test-simpleExchange.js
@@ -1,11 +1,10 @@
-import { test } from './prepare-test-env-ava.js';
-import { E } from '@endo/far';
+import { test } from '../tools/prepare-test-env-ava.js';
+import { E } from '@endo/eventual-send';
 import { setUpZoeForTest } from '@agoric/inter-protocol/test/supports.js';
-import { makeIssuerKit, AmountMath, AssetKind } from '@agoric/ertp';
-import { makeSimpleExchangeAssertions } from './tools/assertions.js';
-import { makeSimpleExchangeHelpers } from './tools/helpers.js';
-import { setupUpgradableSimpleExchange, setupAssets } from './tools/setup.js';
-import { eventLoopIteration } from '@agoric/internal/src/testing-utils.js';
+import { AmountMath, AssetKind, makeIssuerKit } from '@agoric/ertp';
+import { makeSimpleExchangeAssertions } from '../tools/assertions.js';
+import { makeSimpleExchangeHelpers } from '../tools/helpers.js';
+import { setupSimpleExchange, setupAssets } from '../tools/setup.js';
 
 test.beforeEach(async (t) => {
   const { zoe } = await setUpZoeForTest(() => {});
@@ -14,12 +13,14 @@ test.beforeEach(async (t) => {
   const makeSimpleMake = (brand) => (value) => AmountMath.make(brand, value);
   const moola = makeSimpleMake(assets.moolaKit.brand);
   const simoleans = makeSimpleMake(assets.simoleanKit.brand);
+  const nothing = makeSimpleMake(assets.nothingKit.brand);
 
   t.context = {
     zoe,
     assets,
     moola,
     simoleans,
+    nothing,
   };
 });
 
@@ -29,17 +30,14 @@ test('make sell offer', async (t) => {
   const assertions = makeSimpleExchangeAssertions(t);
   const helpers = makeSimpleExchangeHelpers();
 
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(
-    zoe,
-    assets,
-  );
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, assets);
 
   const issuers = await E(zoe).getIssuers(instance);
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
   assertions.assertIssuer(issuers.Price, simoleanKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -53,14 +51,13 @@ test('make sell offer', async (t) => {
   );
 
   const seat = await E(zoe).offer(invitation, sellOrderProposal, sellPayment);
-  await eventLoopIteration();
 
   const offerResult = await E(seat).getOfferResult();
   assertions.assertOfferResult(offerResult, 'Order Added');
 
   expectedBuys = [];
   expectedSells = [sellOrderProposal];
-  orderBook = await E(subscriber).getUpdateSince();
+  orderBook = await E(notifier).getUpdateSince();
   assertions.assertOrderBook(orderBook, expectedBuys, expectedSells);
 });
 
@@ -70,17 +67,14 @@ test('make buy offer', async (t) => {
   const assertions = makeSimpleExchangeAssertions(t);
   const helpers = makeSimpleExchangeHelpers();
 
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(
-    zoe,
-    assets,
-  );
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, assets);
 
   const issuers = await E(zoe).getIssuers(instance);
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
   assertions.assertIssuer(issuers.Price, simoleanKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -90,14 +84,13 @@ test('make buy offer', async (t) => {
   const { buyOrderProposal, buyPayment } = helpers.makeBuyOffer(assets, 3n, 4n);
 
   const seat = await E(zoe).offer(invitation, buyOrderProposal, buyPayment);
-  await eventLoopIteration();
 
   const offerResult = await E(seat).getOfferResult();
   assertions.assertOfferResult(offerResult, 'Order Added');
 
   expectedBuys = [buyOrderProposal];
   expectedSells = [];
-  orderBook = await E(subscriber).getUpdateSince();
+  orderBook = await E(notifier).getUpdateSince();
   assertions.assertOrderBook(orderBook, expectedBuys, expectedSells);
 });
 
@@ -107,17 +100,14 @@ test('make trade', async (t) => {
   const assertions = makeSimpleExchangeAssertions(t);
   const helpers = makeSimpleExchangeHelpers();
 
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(
-    zoe,
-    assets,
-  );
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, assets);
 
   const issuers = await E(zoe).getIssuers(instance);
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
   assertions.assertIssuer(issuers.Price, simoleanKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -139,14 +129,13 @@ test('make trade', async (t) => {
     sellOrderProposal,
     sellPayment,
   );
-  await eventLoopIteration();
 
   const offerResult = await E(aliceSeat).getOfferResult();
   assertions.assertOfferResult(offerResult, 'Order Added');
 
   expectedBuys = [];
   expectedSells = [sellOrderProposal];
-  orderBook = await E(subscriber).getUpdateSince();
+  orderBook = await E(notifier).getUpdateSince();
   assertions.assertOrderBook(orderBook, expectedBuys, expectedSells);
 
   // Bob makes a buy offer
@@ -162,7 +151,6 @@ test('make trade', async (t) => {
     buyOrderProposal,
     buyPayment,
   );
-  await eventLoopIteration();
 
   const bobOfferResult = await E(bobSeat).getOfferResult();
   assertions.assertOfferResult(bobOfferResult, 'Order Added');
@@ -170,7 +158,7 @@ test('make trade', async (t) => {
   // As the trade is made, the order book should be cleared
   expectedBuys = [];
   expectedSells = [];
-  orderBook = await E(subscriber).getUpdateSince();
+  orderBook = await E(notifier).getUpdateSince();
   assertions.assertOrderBook(orderBook, expectedBuys, expectedSells);
 
   // Asset assets are swapped
@@ -190,10 +178,7 @@ test('make offer with wrong issuers', async (t) => {
   const assertions = makeSimpleExchangeAssertions(t);
   const helpers = makeSimpleExchangeHelpers();
 
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(
-    zoe,
-    assets,
-  );
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, assets);
 
   const issuers = await E(zoe).getIssuers(instance);
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
@@ -202,8 +187,8 @@ test('make offer with wrong issuers', async (t) => {
   assertions.assertNotIssuer(issuers.Asset, nothingKit.issuer);
   assertions.assertNotIssuer(issuers.Price, nothingKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -214,20 +199,14 @@ test('make offer with wrong issuers', async (t) => {
     helpers.makeInvalidOffer(assets, 3n, 4n, 'wrongWantIssuer');
 
   let seatPromise = E(zoe).offer(invitation, sellOrderProposal, sellPayment);
-  await eventLoopIteration();
 
   await assertions.assertThrowError(seatPromise, expectedError);
 
   invitation = await E(publicFacet).makeInvitation();
-  ({ sellOrderProposal, sellPayment, expectedError } = helpers.makeInvalidOffer(
-    assets,
-    3n,
-    4n,
-    'wrongGiveIssuer',
-  ));
+  ({ sellOrderProposal, sellPayment, expectedError } =
+    helpers.makeInvalidOffer(assets, 3n, 4n, 'wrongGiveIssuer'));
 
   seatPromise = E(zoe).offer(invitation, sellOrderProposal, sellPayment);
-  await eventLoopIteration();
 
   await assertions.assertThrowError(seatPromise, expectedError);
 });
@@ -238,17 +217,14 @@ test('make offer with offerProposal missing attribute', async (t) => {
   const assertions = makeSimpleExchangeAssertions(t);
   const helpers = makeSimpleExchangeHelpers();
 
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(
-    zoe,
-    assets,
-  );
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, assets);
 
   const issuers = await E(zoe).getIssuers(instance);
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
   assertions.assertIssuer(issuers.Price, simoleanKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -258,19 +234,13 @@ test('make offer with offerProposal missing attribute', async (t) => {
   let { sellOrderProposal, sellPayment, expectedError } =
     helpers.makeInvalidOffer(assets, 3n, 4n, 'missingWant');
   let seatPromise = E(zoe).offer(invitation, sellOrderProposal, sellPayment);
-  await eventLoopIteration();
 
   await assertions.assertThrowError(seatPromise, expectedError);
 
   invitation = await E(publicFacet).makeInvitation();
-  ({ sellOrderProposal, sellPayment, expectedError } = helpers.makeInvalidOffer(
-    assets,
-    3n,
-    4n,
-    'missingGive',
-  ));
+  ({ sellOrderProposal, sellPayment, expectedError } =
+    helpers.makeInvalidOffer(assets, 3n, 4n, 'missingGive'));
   seatPromise = E(zoe).offer(invitation, sellOrderProposal, sellPayment);
-  await eventLoopIteration();
 
   await assertions.assertThrowError(seatPromise, expectedError);
 });
@@ -280,17 +250,14 @@ test('make offer without offerProposal', async (t) => {
   const { moolaKit, simoleanKit } = assets;
   const assertions = makeSimpleExchangeAssertions(t);
 
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(
-    zoe,
-    assets,
-  );
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, assets);
 
   const issuers = await E(zoe).getIssuers(instance);
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
   assertions.assertIssuer(issuers.Price, simoleanKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -299,7 +266,6 @@ test('make offer without offerProposal', async (t) => {
   const invitation = await E(publicFacet).makeInvitation();
 
   const seatPromise = E(zoe).offer(invitation, undefined, undefined);
-  await eventLoopIteration();
 
   await assertions.assertThrowError(
     seatPromise,
@@ -313,17 +279,14 @@ test('offers with null or invalid shapes on the proposals', async (t) => {
   const assertions = makeSimpleExchangeAssertions(t);
   const helpers = makeSimpleExchangeHelpers();
 
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(
-    zoe,
-    assets,
-  );
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, assets);
 
   const issuers = await E(zoe).getIssuers(instance);
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
   assertions.assertIssuer(issuers.Price, simoleanKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -344,7 +307,6 @@ test('offers with null or invalid shapes on the proposals', async (t) => {
         sellOrderProposal,
         sellPayment,
       );
-      await eventLoopIteration();
 
       await assertions.assertThrowError(seatPromise, expectedError);
     },
@@ -362,7 +324,7 @@ test('make offer with NFT', async (t) => {
 
   // Make NFT kit
   const moolaKit = makeIssuerKit('Moola', AssetKind.SET);
-  const { publicFacet, instance } = await setupUpgradableSimpleExchange(zoe, {
+  const { publicFacet, instance } = await setupSimpleExchange(zoe, {
     moolaKit,
     simoleanKit,
   });
@@ -371,8 +333,8 @@ test('make offer with NFT', async (t) => {
   assertions.assertIssuer(issuers.Asset, moolaKit.issuer);
   assertions.assertIssuer(issuers.Price, simoleanKit.issuer);
 
-  const subscriber = await E(publicFacet).getSubscriber();
-  let orderBook = await E(subscriber).getUpdateSince();
+  const notifier = await E(publicFacet).getNotifier();
+  let orderBook = await E(notifier).getUpdateSince();
 
   let expectedBuys = [];
   let expectedSells = [];
@@ -394,14 +356,13 @@ test('make offer with NFT', async (t) => {
     sellOrderProposal,
     sellPayment,
   );
-  await eventLoopIteration();
 
   const aliceOfferResult = await E(aliceSeat).getOfferResult();
   assertions.assertOfferResult(aliceOfferResult, 'Order Added');
 
   expectedBuys = [];
   expectedSells = [sellOrderProposal];
-  orderBook = await E(subscriber).getUpdateSince();
+  orderBook = await E(notifier).getUpdateSince();
   assertions.assertOrderBook(orderBook, expectedBuys, expectedSells);
 
   // Bob makes a buy offer
@@ -417,7 +378,6 @@ test('make offer with NFT', async (t) => {
     buyOrderProposal,
     buyPayment,
   );
-  await eventLoopIteration();
 
   const bobOfferResult = await E(bobSeat).getOfferResult();
   assertions.assertOfferResult(bobOfferResult, 'Order Added');
@@ -425,7 +385,7 @@ test('make offer with NFT', async (t) => {
   // As the trade is made, the order book should be cleared
   expectedBuys = [];
   expectedSells = [];
-  orderBook = await E(subscriber).getUpdateSince();
+  orderBook = await E(notifier).getUpdateSince();
   assertions.assertOrderBook(orderBook, expectedBuys, expectedSells);
 
   // Asset assets are swapped
